chore(gateway): rename protocol module to gateway
